Title : Protecting Video Routes

+++ controllers/videoController.js
@@ -71,11 +71,15 @@ export const videoDetail = async (req, res) => {
 
 export const getEditVideo = async (req, res) => {
   const {
-    params: {id}
+    params: { id }
   } = req;
-  try{
+  try {
     const video = await Video.findById(id); // findById 가 뭔지 알아내기
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== String(req.user.id)) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -83,11 +87,11 @@ export const getEditVideo = async (req, res) => {
 // Video를 업데이트
 export const postEditVideo = async (req, res) => {
   const {
-    params: {id},
-    body: {title,  description}
+    params: { id },
+    body: { title, description }
   } = req;
   try {
-    await Video.findOneAndUpdate({_id: id}, {title,  description}); // id를 찾고, 제목과 설명을 업데이트한다.
+    await Video.findOneAndUpdate({ _id: id }, { title, description }); // id를 찾고, 제목과 설명을 업데이트한다.
     res.redirect(routes.videoDetail(id));
   } catch (error) {
     res.redirect(routes.home);
@@ -98,10 +102,15 @@ export const postEditVideo = async (req, res) => {
 
 export const deleteVideo = async(req, res) => {
   const {
-    params: {id}
+    params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (String(video.creator) !== String(req.user.id)) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
